[MIDAS] support visual studio

+++ src/main/cc/sample.cc
@@ -6,7 +6,7 @@
 #ifdef ENABLE_SNAPSHOT
 std::array<std::vector<std::string>, CHAIN_NUM> sample_t::signals = {};
 std::array<std::vector<size_t>,      CHAIN_NUM> sample_t::widths  = {};
-std::array<std::vector<ssize_t>,     CHAIN_NUM> sample_t::depths = {};
+std::array<std::vector<int>,         CHAIN_NUM> sample_t::depths = {};
 size_t sample_t::chain_len[CHAIN_NUM] = {0};
 size_t sample_t::chain_loop[CHAIN_NUM] = {0};
 
@@ -47,7 +47,7 @@ std::ostream& dump_s(std::ostream &os,
 void sample_t::init_chains(std::string filename) {
   std::fill(signals.begin(), signals.end(), std::vector<std::string>());
   std::fill(widths.begin(),  widths.end(),  std::vector<size_t>());
-  std::fill(depths.begin(), depths.end(), std::vector<ssize_t>());
+  std::fill(depths.begin(), depths.end(), std::vector<int>());
   std::ifstream file(filename.c_str());
   if (!file) {
     fprintf(stderr, "Cannot open %s\n", filename.c_str());
@@ -60,7 +60,7 @@ void sample_t::init_chains(std::string filename) {
     std::string signal;
     iss >> type >> signal;
     size_t width;
-    ssize_t depth;
+    int depth;
     iss >> width >> depth;
     if (signal == "null") signal = "";
     signals[type].push_back(signal);
@@ -120,12 +120,12 @@ size_t sample_t::read_chain(CHAIN_TYPE type, const char* snap, size_t start) {
   size_t t = static_cast<size_t>(type);
   std::vector<std::string> &chain_signals = signals[t];
   std::vector<size_t> &chain_widths = widths[t];
-  std::vector<ssize_t> &chain_depths = depths[t];
+  std::vector<int> &chain_depths = depths[t];
   for (size_t i = 0 ; i < chain_loop[type] ; i++) {
     for (size_t s = 0 ; s < chain_signals.size() ; s++) {
       std::string &signal = chain_signals[s];
       size_t width = chain_widths[s];
-      ssize_t depth = chain_depths[s];
+      int depth = chain_depths[s];
       if (!signal.empty()) {
         char substr[1025];
         if (width > 1024) throw std::out_of_range("width should be <= 1024");
@@ -143,7 +143,7 @@ size_t sample_t::read_chain(CHAIN_TYPE type, const char* snap, size_t start) {
             add_cmd(new load_t(type, s, data, value.get_size(), -1));
             break;
           case SRAM_CHAIN:
-            if (static_cast<ssize_t>(i) < depth)
+            if (static_cast<int>(i) < depth)
               add_cmd(new load_t(type, s, data, value.get_size(), i));
             break;
           case CNTR_CHAIN:
@@ -172,7 +172,7 @@ void sample_t::add_force(force_t* f) {
 }
 
 void sample_t::dump_forces() {
-  for (ssize_t i = force_bins.size() - 1 ; i >= 0 ; i--) {
+  for (int i = force_bins.size() - 1 ; i >= 0 ; i--) {
     std::vector<force_t*> force_bin = force_bins[i];
     for (size_t k = 0 ; k < force_bin.size() ; k++) {
       cmds.push_back(force_bin[k]);
